requirements.txt
dataclasses
openai
transformers
faiss-cpu
langchain
pydantic
python-dotenv
typing_extensions